+++ tests/amm/amm.ts
@@ -1,0 +1,33 @@
+import { amm, payer } from "../helpers";
+import { PublicKey } from "@solana/web3.js";
+
+/**
+ * This is a call to the dev version of this endpoint. Due to the way the anchor
+ * loads programs on localnet (so that we can use any pubkey and don't have to
+ * sign the program deploy), the programs on localnet don't have the same
+ * structure in terms of having a data account as with normal deployment.
+ */
+export async function createProgramToll(
+  programTollAuthority: PublicKey = payer.publicKey
+) {
+  const programToll = programTollAddress();
+
+  await amm.methods
+    .createProgramToll()
+    .accounts({
+      programTollAuthority,
+      programToll,
+    })
+    .signers([payer])
+    .rpc();
+
+  return programToll;
+}
+
+export function programTollAddress(): PublicKey {
+  const [programToll, _bumpSeed] = PublicKey.findProgramAddressSync(
+    [Buffer.from("toll")],
+    amm.programId
+  );
+  return programToll;
+}